Include a recipe in the default.
The private recipe, `_message.rb`, defines the `company` attribute.
Because the recipe `default.rb`, also defines the `company` attribute,
these definitions conflict. Because the have the same precedence,
"default," the **last one wins** and so the value of the `company`
attribute after the converge operation is the value defined in the file
`_message.rb`.